+++ src/propedit/tag_target.cpp
@@ -243,7 +243,7 @@ tag_target_c::read_segment_info_and_tracks() {
   m_timestamp_scale = segment_info ? FindChildValue<KaxTimecodeScale>(*segment_info, 1000000ull) : 1000000ull;
 
   if (tracks && dynamic_cast<KaxTracks *>(tracks.get())) {
-    for (size_t idx = 0, num_children = tracks->ListSize(); idx < num_children; ++idx) {
+    for (int idx = 0, num_children = tracks->ListSize(); idx < num_children; ++idx) {
       auto track = dynamic_cast<KaxTrackEntry *>((*tracks)[idx]);
 
       if (!track)
@@ -284,7 +284,7 @@ tag_target_c::account_block_group(KaxBlockGroup &block_group,
   auto frame_duration  = block_duration ? static_cast<uint64_t>(block_duration->GetValue() * m_timestamp_scale / num_frames) : m_default_durations_by_number[block->TrackNum()];
   auto first_timestamp = block->GlobalTimecode();
 
-  for (size_t idx = 0; idx < num_frames; ++idx) {
+  for (int idx = 0; idx < static_cast<int>(num_frames); ++idx) {
     auto &data_buffer = block->GetBuffer(idx);
     stats_itr->second.account(first_timestamp + idx * frame_duration, frame_duration, data_buffer.Size());
   }
@@ -304,7 +304,7 @@ tag_target_c::account_simple_block(KaxSimpleBlock &simple_block,
   auto frame_duration  = m_default_durations_by_number[simple_block.TrackNum()];
   auto first_timestamp = simple_block.GlobalTimecode();
 
-  for (size_t idx = 0; idx < num_frames; ++idx) {
+  for (int idx = 0; idx < static_cast<int>(num_frames); ++idx) {
     auto &data_buffer = simple_block.GetBuffer(idx);
     stats_itr->second.account(first_timestamp + idx * frame_duration, frame_duration, data_buffer.Size());
   }
@@ -312,7 +312,7 @@ tag_target_c::account_simple_block(KaxSimpleBlock &simple_block,
 
 void
 tag_target_c::account_one_cluster(KaxCluster &cluster) {
-  for (size_t idx = 0, num_children = cluster.ListSize(); idx < num_children; ++idx) {
+  for (int idx = 0, num_children = cluster.ListSize(); idx < num_children; ++idx) {
     auto child = cluster[idx];
 
     if (Is<KaxBlockGroup>(child))
